Add modify and delete reesources by ownership requirement
Added Resource operation requirement and handler which checks if create/get then allow or rest of verbs equals check ownership. Changed GameCompany entity to store owner id (added migration). Changed GameCompanyService's delete and update to send to injected authorization service user, resource and requirement for validation

+++ VideoGamesCompaniesAPI/Services/GameCompanyService.cs
@@ -1,11 +1,14 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
+using VideoGamesCompaniesAPI.Authorization;
 using VideoGamesCompaniesAPI.Entities;
 using VideoGamesCompaniesAPI.Exceptions;
 using VideoGamesCompaniesAPI.Models;
@@ -14,11 +17,11 @@ namespace VideoGamesCompaniesAPI.Services
 {
     public interface IGameCompanyService
     {
-        int Create(CreateGameCompanyDto dto);
-        void Delete(int id);
+        int Create(CreateGameCompanyDto dto, int userId);
+        void Delete(int id, ClaimsPrincipal user);
         IEnumerable<GameCompanyDto> GetAll();
         GameCompanyDto GetById(int id);
-        void Update(int id, UpdateGameCompanyDto dto);
+        void Update(int id, UpdateGameCompanyDto dto, ClaimsPrincipal user);
     }
 
     public class GameCompanyService : IGameCompanyService
@@ -26,22 +29,33 @@ namespace VideoGamesCompaniesAPI.Services
         private readonly GameCompanyDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<GameCompanyService> _logger;
+        private readonly IAuthorizationService _authorizationService;
 
-        public GameCompanyService(GameCompanyDbContext dbContext, IMapper mapper, ILogger<GameCompanyService> logger)
+        public GameCompanyService(GameCompanyDbContext dbContext, IMapper mapper, ILogger<GameCompanyService> logger, IAuthorizationService authorizationService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
         }
 
-        public void Update(int id, UpdateGameCompanyDto dto)
+        public void Update(int id, UpdateGameCompanyDto dto, ClaimsPrincipal user)
         {
+
             var gameCompany = _dbContext
                 .GameCompanies
                 .FirstOrDefault(gc => gc.Id == id);
 
             if (gameCompany is null)
                 throw new NotFoundException("Game company not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, gameCompany,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             gameCompany.Name = dto.Name;
             gameCompany.Description = dto.Description;
@@ -78,16 +92,17 @@ namespace VideoGamesCompaniesAPI.Services
             return results;
         }
 
-        public int Create(CreateGameCompanyDto dto)
+        public int Create(CreateGameCompanyDto dto, int userId)
         {
             var gameCompany = _mapper.Map<GameCompany>(dto);
+            gameCompany.CreatedById = userId;
             _dbContext.GameCompanies.Add(gameCompany);
             _dbContext.SaveChanges();
 
             return gameCompany.Id;
         }
 
-        public void Delete(int id)
+        public void Delete(int id, ClaimsPrincipal user)
         {
             var gameCompany = _dbContext
                 .GameCompanies
@@ -96,6 +111,13 @@ namespace VideoGamesCompaniesAPI.Services
             if (gameCompany is null)
                 throw new NotFoundException("Game company not found");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, gameCompany,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _dbContext.GameCompanies.Remove(gameCompany);
             _dbContext.SaveChanges();
